[CDEC-416] Sinbin `Pos.Infra.Communication.Protocol`
This module had a number of dependencies which needed to be sinbinned
with it.
